feat: modularize feature of caching input
