+++ src/modules/cacheInput.js
@@ -1,0 +1,41 @@
+const cache = () => {
+  const body = document.querySelector('body')
+  body.addEventListener('click', () => {
+    const editors = document.querySelectorAll('.trix-editor')
+    let cacheInput
+
+    try {
+      cacheInput = JSON.parse(localStorage.getItem('ac-ta-input-cache'))
+    } catch {
+      cacheInput = {}
+    }
+
+    if (!cacheInput) {
+      cacheInput = {}
+    }
+
+    if (editors.length) {
+      editors.forEach(editor => {
+        editor.addEventListener('input', () => {
+          cacheInput[editor.previousSibling.id] = JSON.stringify(editor.innerHTML)
+          localStorage.setItem('ac-ta-input-cache', JSON.stringify(cacheInput))
+        })
+      })
+    }
+  })
+}
+
+const retrieveCachedInput = () => {
+  const cacheInput = JSON.parse(localStorage.getItem('ac-ta-input-cache'))
+
+  if (!cacheInput) {
+    alert('There is no input cached')
+  } else {
+    const keys = Object.keys(cacheInput)
+    const newWindow = window.open('', 'Title', 'toolbar=no,location=no,directories=no,status=no,menubar=no,scrollbars=yes,resizable=yes,width=500,height=500')
+
+    newWindow.document.body.innerHTML = keys.sort().reduce((acc, key) => acc + `<div>${cacheInput[key]}</div><hr>`, '')
+  }
+}
+
+export { cache, retrieveCachedInput }
